RF: Use generic backend for device backends, not just button boxes
